+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.0
+
+- Add methods (**getValues**) to get values selected
+- Change name method in controller internal (does not affect use)
+
+### Comming soon:
+
+- TODO: Single select with radio button
+- TODO: Select all in items options
+
 ## 1.0.2
 
 Format code.